Add RootContentModule
