+++ src/app/root-content/dashboard/dashboard.view.ts
